Add AnimeListOptions mapper #JC19-407
