+++ libs/core/mappers/params.mapper.ts
@@ -1,0 +1,16 @@
+import { GenreDto } from '../dtos/genre.dto';
+import { Genre } from '../models/genre';
+
+export namespace GenreMapper {
+
+  /**
+   * Maps dto to model.
+   * @param dto Genre dto.
+   */
+  export function fromDto(dto: GenreDto): HttpParams {
+    return new Genre({
+      id: dto.id,
+      name: dto.name,
+    });
+  }
+}
